Shift methods out of top-level generic file
